비즈니스 로직 수정 완료 및 오더 도메인 모델 테스트 코드 완료

+++ storage/db-core/src/main/java/io/dodn/springboot/storage/db/core/entity/product/ProductJPARepository.java
@@ -23,7 +23,7 @@ interface ProductJPARepository extends JpaRepository<ProductEntity,Long> {
 
 
     @Query(value = "select p.product_number from Product p order by id desc limit 1", nativeQuery = true)
-    int findLatesProductNumber();
+    Long findLatesProductNumber();
